move isListOrPath predicates
