+++ frontends/ol-search-ui/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./components"
 export * from "./interfaces"
+export { isStaffListOrPath, isUserListOrPath } from "./util"
