+++ frontends/ol-search-ui/src/util.tsx
@@ -4,6 +4,8 @@ import {
   CoursePrice,
   CourseInstructor,
   LearningResourceType as LR,
+  UserList,
+  StaffList,
   TYPE_FAVORITES
 } from "./interfaces"
 import React, { useState, useEffect } from "react"
@@ -235,3 +237,22 @@ const formatPrice = (price: number | null | undefined): string => {
 
 export const absolutizeURL = (url: string) =>
   new URL(url, window.location.origin).toString()
+
+const isUserListOrPath = (
+  resource: UserList | StaffList
+): resource is UserList => {
+  return (
+    resource.object_type === LR.Userlist ||
+    resource.object_type === LR.LearningPath
+  )
+}
+const isStaffListOrPath = (
+  resource: UserList | StaffList
+): resource is StaffList => {
+  return (
+    resource.object_type === LR.StaffList ||
+    resource.object_type === LR.StaffPath
+  )
+}
+
+export { isUserListOrPath, isStaffListOrPath }
